Move friend endpoint to a seperate file
